Fix bug when formatting TCRYPT volumes
When formatting a TCRYPT volume without overwriting the data, the
chi-squared test will still pass, so the block device will still have
type "crypto_unknown" after filesystem signatures were wiped. This
causes the subsequent wait_for_filesystem() to time out, because the
block never has an empty type.

This commit makes wait_for_filesystem() also return if the type is
"crypto_unknown", when it is waiting for an empty type.